Added a User00 keycode that toggles the new RGB
The new RGB mode is disabled by default but can be enabled using the new user key code. Toggling a mode change via a key (RGB_MOD) will unset the toggle. Using the GUI can cause confusion as to what mode is the current so ensure the new mode is turned off on the keyboard before setting one in the GUI

Also moved the firmware to a vial folder as the vial docs indicate it should be.